#15 - Removed config.ts in core as types were moved to types.ts (for clearer initialization procedure)
