+++ packages/core/src/node/node.ts
@@ -3,8 +3,7 @@ import { createHelia, Helia } from "helia";
 import { createLibp2p, Libp2p } from "libp2p";
 import { MemoryBlockstore } from "blockstore-core";
 import { LevelBlockstore } from "blockstore-level";
-import { OrbitDB } from "@orbitdb/core";
-import { OiConfigDatabase, OiConfigHelia } from "../config";
+import { OiConfigDatabase, OiConfigHelia } from "../types";
 import { NodeLibp2pConfig } from './libp2p';
 import { monitorPeers } from "./peers";
 import { OiLoggerInterface } from "../types";
